Some notes for class abstraction and the graph class.

+++ lectures/lec13/src/Bank_account.cpp
@@ -14,7 +14,7 @@ const std::string& Bank_account::owner() const
     return owner_;
 }
 
-unsigned int Bank_account::balance() const
+unsigned long Bank_account::balance() const
 {
     return balance_;
 }
@@ -24,13 +24,13 @@ void Bank_account::change_owner(const std::string& s)
     owner_ = s;
 }
 
-void Bank_account::deposit(unsigned int amount)
+void Bank_account::deposit(unsigned long amount)
 {
     // This can overflow!
     balance_ += amount;
 }
 
-bool Bank_account::withdraw(unsigned int amount)
+bool Bank_account::withdraw(unsigned long amount)
 {
     if (amount <= balance_) {
         balance_ -= amount;
