+++ lectures/lec13/src/Bank_account.h
@@ -13,18 +13,18 @@ public:
     // Returns the name of the account owner.
     const std::string& owner() const;
     // Returns the balance of the account.
-    unsigned int balance() const;
+    unsigned long balance() const;
 
     // Assigns the account a different owner.
     void change_owner(const std::string& new_owner);
     // Deposits the given quantity in the account.
-    void deposit(unsigned int);
+    void deposit(unsigned long);
     // Attempts to withdraw the given quantity from the account; returns
     // `true` if the withdrawal succeeds and `false` otherwise.
-    bool withdraw(unsigned int);
+    bool withdraw(unsigned long);
 
 private:
     unsigned long id_;
     std::string owner_;
-    unsigned int balance_;
+    unsigned long balance_;
 };
